Fixed some sprite/texture issues. Beginning work on goClasses.
goClass = gameobjectClass. All goClasses inherit from goVirtual. This way all that needs to be edited in the goClass itself is its behavior methods. Structural methods for loading behaviors in should be handled by goVirtual. We can then declare a GameObj<goClass> to interface with Scenes, which then interfaces with GameRunner.
So much abstraction.